remove unconstrained strategy, rename cooling to unconstrained
