+++ mt-kahypar/partition/context_enum_classes.cpp
@@ -230,8 +230,6 @@ namespace mt_kahypar {
     switch (algo) {
       case FMAlgorithm::kway_fm: return os << "kway_fm";
       case FMAlgorithm::unconstrained: return os << "unconstrained";
-      case FMAlgorithm::combined: return os << "combined";
-      case FMAlgorithm::cooling: return os << "cooling";
       case FMAlgorithm::do_nothing: return os << "fm_do_nothing";
         // omit default case to trigger compiler warning for missing cases
     }
@@ -451,10 +449,6 @@ namespace mt_kahypar {
       return FMAlgorithm::kway_fm;
     } else if (type == "unconstrained") {
       return FMAlgorithm::unconstrained;
-    } else if (type == "combined") {
-      return FMAlgorithm::combined;
-    } else if (type == "cooling") {
-      return FMAlgorithm::cooling;
     } else if (type == "do_nothing") {
       return FMAlgorithm::do_nothing;
     }
