+++ mt-kahypar/partition/context_enum_classes.h
@@ -156,8 +156,6 @@ enum class LabelPropagationAlgorithm : uint8_t {
 enum class FMAlgorithm : uint8_t {
   kway_fm,
   unconstrained,
-  combined,
-  cooling,
   do_nothing
 };
