+++ mt-kahypar/partition/factories.h
@@ -50,7 +50,6 @@
 #include "mt-kahypar/partition/refinement/fm/strategies/i_fm_strategy.h"
 #include "mt-kahypar/partition/refinement/fm/strategies/gain_cache_strategy.h"
 #include "mt-kahypar/partition/refinement/fm/strategies/unconstrained_strategy.h"
-#include "mt-kahypar/partition/refinement/fm/strategies/cooling_strategy.h"
 #include "mt-kahypar/partition/refinement/gains/gain_definitions.h"
 #include "mt-kahypar/partition/refinement/flows/scheduler.h"
 #include "mt-kahypar/partition/refinement/flows/flow_refiner.h"
@@ -104,7 +103,6 @@ using DefaultFMDispatcher = kahypar::meta::StaticMultiDispatchFactory<
                             kahypar::meta::Typelist<TypeTraitsList, GainTypes>>;
 
 using UnconstrainedFMDispatcher = DefaultFMDispatcher;
-using CoolingFMDispatcher = DefaultFMDispatcher;
 
 using FMStrategyFactory = kahypar::meta::Factory<FMAlgorithm, IFMStrategy* (*)(const Context&, FMSharedData&)>;
 
@@ -117,11 +115,6 @@ using UnconstrainedFMStrategyDispatcher = kahypar::meta::StaticMultiDispatchFact
                                           UnconstrainedStrategy,
                                           IFMStrategy,
                                           kahypar::meta::Typelist<TypeTraitsList, GainTypes>>;
-
-using CoolingFMStrategyDispatcher = kahypar::meta::StaticMultiDispatchFactory<
-                                    CoolingStrategy,
-                                    IFMStrategy,
-                                    kahypar::meta::Typelist<TypeTraitsList, GainTypes>>;
 
 using FlowSchedulerFactory = kahypar::meta::Factory<FlowAlgorithm,
                               IRefiner* (*)(const HypernodeID, const HyperedgeID, const Context&, gain_cache_t)>;
